+++ src/framework/renderer/fx/material/material_fx.h
@@ -1,5 +1,5 @@
-#ifndef VKFRAMEWORK_RENDERER_FX_MATERIAL_FX_H_
-#define VKFRAMEWORK_RENDERER_FX_MATERIAL_FX_H_
+#ifndef VKFRAMEWORK_RENDERER_FX_MATERIAL_MATERIAL_FX_H_
+#define VKFRAMEWORK_RENDERER_FX_MATERIAL_MATERIAL_FX_H_
 
 #include "framework/common.h"
 #include "framework/renderer/renderer.h"
@@ -287,4 +287,4 @@ class TMaterialFx : public MaterialFx {
 
 /* -------------------------------------------------------------------------- */
 
-#endif // VKFRAMEWORK_RENDERER_FX_MATERIAL_FX_H_
+#endif // VKFRAMEWORK_RENDERER_FX_MATERIAL_MATERIAL_FX_H_
