+++ src/framework/scene/private/gltf_loader.cc
@@ -13,8 +13,8 @@ static constexpr bool kFrameworkHasDraco{false};
 
 #include "framework/renderer/sampler_pool.h"
 
-#include "framework/renderer/fx/material/pbr_metallic_roughness.h" //
-#include "framework/renderer/fx/material/unlit.h" //
+#include "framework/renderer/fx/material/impl/pbr_metallic_roughness.h" //
+#include "framework/renderer/fx/material/impl/unlit.h" //
 
 /* -------------------------------------------------------------------------- */
