+++ src/samples/09_post_process/main.cc
@@ -16,7 +16,7 @@
 #include "framework/renderer/fx/postprocess/fragment/impl/normaldepth_edge.h"
 #include "framework/renderer/fx/postprocess/fragment/impl/object_edge.h"
 
-#include "framework/renderer/fx/material/pbr_metallic_roughness.h"
+#include "framework/renderer/fx/material/impl/pbr_metallic_roughness.h"
 
 namespace shader_interop {
 #include "shaders/interop.h"
